\name{BiocGenerics-package}

\alias{BiocGenerics-package}
\alias{BiocGenerics}
\docType{package}

\title{Generic functions for Bioconductor}

\description{
  Generic functions that are needed by many other packages.
}

\details{
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}

\author{The Bioconductor Dev Team}

\seealso{
}

\keyword{package}
